Added receivier, improved register optimization
Receiver component and appears to work when connected to the output
of the transmitter component (currently tested only in the top-level
COMPY-V testbench).

Also set default value on expected start bits to '1' so the register can
be automatically optimized out in designs where the number of start bits
cannot be set (i.e. *all* UART implementations).